Actualizando, mejorando y optimizando los test de controladores para las entidades Doctor y Patient usando TestRestTemplate y MockMCV (ambos quedaron funcionando perfectamente).

+++ src/main/java/com/juan/estevez/app/dto/DoctorDTO.java
@@ -11,9 +11,9 @@ public class DoctorDTO implements Serializable {
 	private String doctorsName;
 	private String idType;
 	private String numberProfessionalCard;
-	private double yearsExperience;
+	private int yearsExperience;
 	private String specialty;
-	private int attentionStartTime;
-	private int attentionEndTime;
+	private double attentionStartTime;
+	private double attentionEndTime;
 	
 }
